+++ src/main/java/com/placki/companyresources/dto/CreateResourceResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
@@ -15,6 +16,6 @@ import java.time.LocalDate;
 public class CreateResourceResponse {
     private String name;
     private LocalDate bookingDate;
-    private double priceUSD;
-    private double pricePLN;
+    private BigDecimal priceUSD; // Changed from double to BigDecimal
+    private BigDecimal pricePLN; // Changed from double to BigDecimal
 }
