+++ src/main/java/com/placki/companyresources/dto/NbpRateResponse.java
@@ -2,6 +2,7 @@ package com.placki.companyresources.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 @Data
 public class NbpRateResponse {
@@ -13,6 +14,6 @@ public class NbpRateResponse {
     public static class Rate {
         private String no;
         private String effectiveDate;
-        private double mid;
+        private BigDecimal mid; // Changed from double to BigDecimal
     }
 }
